pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polymarket-weather-trade-bot"
version = "0.1.0"
description = "Automated Polymarket weather temperature trading bot"
requires-python = ">=3.11"
dependencies = [
    "py-clob-client>=0.5.0",
    "httpx>=0.27.0",
    "apscheduler>=3.10.0",
    "aiosqlite>=0.20.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "flask>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
